+++ src/main/java/seedu/address/logic/commands/contacts/ContactListCommand.java
@@ -1,14 +1,16 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.contacts;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
 
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 
 /**
  * Lists all persons in the address book to the user.
  */
-public class ListCommand extends Command {
+public class ContactListCommand extends Command {
 
     public static final String COMMAND_WORD = "list";
